fix & cleanup
