+++ src/main/java/com/alexyekymov/employeemanager/repository/EmployeeRepository.java
@@ -1,0 +1,12 @@
+package com.alexyekymov.employeemanager.repository;
+
+import com.alexyekymov.employeemanager.model.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface EmployeeRepository extends JpaRepository<Employee,Long> {
+    void deleteEmployeeById(Long id);
+
+    Optional<Employee> findEmployeeById(Long id);
+}
